JAMES-2008 Solve various JPA troubbles
