+++ backends-common/jpa/src/test/java/org/apache/james/backends/jpa/JpaTestCluster.java
@@ -41,7 +41,7 @@ public class JpaTestCluster {
     public static JpaTestCluster create(List<Class<?>> clazz) {
         HashMap<String, String> properties = new HashMap<String, String>();
         properties.put("openjpa.ConnectionDriverName", org.h2.Driver.class.getName());
-        properties.put("openjpa.ConnectionURL", "jdbc:h2:mem:mailboxintegration;DB_CLOSE_DELAY=-1"); // Memory H2 database
+        properties.put("openjpa.ConnectionURL", "jdbc:h2:mem:mailboxintegrationtesting;DB_CLOSE_DELAY=-1"); // Memory H2 database
         properties.put("openjpa.jdbc.SynchronizeMappings", "buildSchema(ForeignKeys=true)"); // Create Foreign Keys
         properties.put("openjpa.jdbc.MappingDefaults", "ForeignKeyDeleteAction=restrict, JoinForeignKeyDeleteAction=restrict");
         properties.put("openjpa.jdbc.SchemaFactory", "native(ForeignKeys=true)");
@@ -84,7 +84,7 @@ public class JpaTestCluster {
         EntityManager entityManager = entityManagerFactory.createEntityManager();
         entityManager.getTransaction().begin();
         for(String tableName: tables) {
-            entityManager.createNativeQuery("TRUNCATE table " + tableName).executeUpdate();
+            entityManager.createNativeQuery("DELETE FROM " + tableName).executeUpdate();
         }
         entityManager.getTransaction().commit();
         entityManager.close();
